Decouple starting of JournalNodes and NameNodes
In HA setup, the JournalNodes MUST be started before the NameNodes.
Previously, the starting tasks were coupled both ways (i.e. if an
operator targeted just the NameNodes for starting, the JournalNodes
would be started first, which is correct and desirable.  However, if an
operator targeted just the JournalNodes for starting, that would also
start the NameNodes.  There is no reason for the second option).  Make
it so that starting the NameNodes will automatically start the
JournalNodes, but the JournalNodes can be started independently of the
NameNodes if one so chooses.