Add environment file for Hadoop daemons
The initscripts for the various Hadoop daemons source a file in
`/etc/default` that wasn't being included by the systemd service units.
The lack of that file caused breakages (e.g. not determining the correct
libexec directory).  Add that file, add a task to distribute it, and add
the ability to source it to the systemd units.