Remove domain name from hdfs-site setting
The config key `dfs.ha.namenodes.mycluster` will cause starting the
NameNode to fail if the domain name is present.  It will try to use the
nameservice ID as the hostname of the node because of an "unresolvable"
address.  Remove it.  Also, move `format_hdfs` to `group_vars` instead
of setting the fact in the `start-hdfs.yml` playbook.