Fix ssh key distribution for ssh fencing
Use of the `synchronize` module would require some knowledge about a
particular site to make it work.  Change the method of of key
distribution so that each NameNode generates its own keypair and
distribute the public keys that way.  Also add host keys to avoid
prompting in the case of a fencing action.