Update defaults files for HDFS
While testing with a Kerberized HDFS cluster, it was discovered that
there were settings that were necessary not only in
`/etc/default/hadoop`, but also `/etc/default/hadoop-hdfs-$daemon`.  Add
this file as a template, add vars to manage the settings in this file
for flexibility, and add a task to distribute this file to all nodes
that will be running an HDFS daemon.  Also, remove redundant settings
from `hadoop-env.sh`.