Allow situational enabling/disabling of services
This PR is rooted in takeaways from production.  Previously, it was
considered undesirable for services to be enabled (that is,
automatically started on boot).  This is due to the nature of Hadoop;
services seem to need to be started in order.  However, in production,
if a node goes down, we don't want to have to go and start all services
that are supposed to be running on that node manually when we recover
it.  Thus, the shadow services (DataNodes and NodeManagers) should be
enabled so that they come back up if something happens to them during
production.