Remove `export` from hdfs role env template
The file `/etc/default/hadoop-hdfs-$daemon` is meant to be read in by
systemd, not executed by a shell.  As such, the `export` commands there
either cause the mechanism to break (at worst) or don't add any value
(at best).  Remove the `export`s and write the file as text.